Cleaning up display

+++ FightingGameWithN-Gram/Assets/Scripts/Character/ChanceDisplay.cs
@@ -20,9 +20,9 @@ public class ChanceDisplay : MonoBehaviour
             ActionChance chances = new ActionChance();
             chances = N_Gram.calculateNextPick(target_character.actionLog);
             //display
-            LinkedText.text += "Attack : " + chances.Attack*100 + "%\n";
-            LinkedText.text += "Block : " + chances.Block*100 + "%\n";
-            LinkedText.text += "Throw : " + chances.Throw*100 + "%\n";
+            LinkedText.text += "Attack : " + (int)(chances.Attack*100) + "%\n";
+            LinkedText.text += "Block : " + (int)(chances.Block*100) + "%\n";
+            LinkedText.text += "Throw : " + (int)(chances.Throw*100) + "%\n";
         }
     }
